Refactor renderer structure.
