+++ src/rust/lib/graph-editor/src/component/visualization/renderer/sample.rs
@@ -1,6 +1,4 @@
 //! The `sample` modules contains some examples of visualisations.
 //! TODO further describe the examples, how they work and how to extend them.
-//!
-mod bubble_chart;
 
-pub use bubble_chart::*;
+pub mod native;
